requirements.txt
Flask
strands-agents
youtube-search-python
yt-dlp==3.0.0
